remove fee summary from App.vue add ViewWraper instead
